backend/requirements.txt
mcp>=1.5.0
fastapi>=0.104.0
uvicorn>=0.23.2
httpx>=0.25.0
python-dotenv>=1.0.0
starlette>=0.27.0
anyio>=3.7.1
google-generativeai>=0.3.0
Pillow>=10.0.0
